requirements.txt
numpy<2.0
matplotlib==3.7.1
pymongo==4.3.3
pydantic==2.0.0
requests==2.28.2
python-dotenv==1.0.0
